alternate event handler splitting approach
