+++ run/client/options.txt
@@ -1,4 +1,4 @@
-version:2865
+version:2975
 autoJump:false
 autoSuggestions:true
 chatColors:true
@@ -67,6 +67,7 @@ mouseWheelSensitivity:1.0
 rawMouseInput:true
 glDebugVerbosity:1
 skipMultiplayerWarning:true
+skipRealms32bitWarning:false
 hideMatchedNames:true
 joinedFirstServer:true
 hideBundleTutorial:false
@@ -107,15 +108,8 @@ key_key.hotbar.6:key.keyboard.6
 key_key.hotbar.7:key.keyboard.7
 key_key.hotbar.8:key.keyboard.8
 key_key.hotbar.9:key.keyboard.9
-key_key.toggleLiquids:key.keyboard.unknown
-key_key.toggleVisible:key.keyboard.unknown
-key_key.curios.open.desc:key.keyboard.g
-key_key.arsmagicalegacy.next_shape_group:key.keyboard.c:SHIFT
-key_key.arsmagicalegacy.configure_spell:key.keyboard.m:CONTROL
-key_key.configured.open_mod_list:key.keyboard.unknown
 key_key.jei.toggleOverlay:key.keyboard.o:CONTROL
 key_key.jei.focusSearch:key.keyboard.f:CONTROL
-key_key.jei.toggleEditMode:key.keyboard.unknown
 key_key.jei.previousPage:key.keyboard.page.up
 key_key.jei.nextPage:key.keyboard.page.down
 key_key.jei.toggleBookmarkOverlay:key.keyboard.unknown
@@ -129,13 +123,24 @@ key_key.jei.previousSearch:key.keyboard.up
 key_key.jei.nextSearch:key.keyboard.down
 key_key.jei.toggleCheatMode:key.keyboard.unknown
 key_key.jei.cheatOneItem:key.mouse.left
+key_key.jei.cheatOneItem2:key.mouse.right
 key_key.jei.cheatItemStack:key.mouse.left:SHIFT
 key_key.jei.cheatItemStack2:key.mouse.middle
+key_key.jei.toggleCheatModeConfigButton:key.mouse.left:CONTROL
+key_key.jei.toggleEditMode:key.keyboard.unknown
+key_key.jei.toggleHideIngredient:key.mouse.left:CONTROL
+key_key.jei.toggleWildcardHideIngredient:key.mouse.right:CONTROL
 key_key.jei.recipeBack:key.keyboard.backspace
 key_key.jei.previousRecipePage:key.keyboard.page.up
 key_key.jei.nextRecipePage:key.keyboard.page.down
 key_key.jei.previousCategory:key.keyboard.page.up:SHIFT
 key_key.jei.nextCategory:key.keyboard.page.down:SHIFT
+key_key.jei.copy.recipe.id:key.keyboard.unknown
+key_key.toggleLiquids:key.keyboard.unknown
+key_key.toggleVisible:key.keyboard.unknown
+key_key.curios.open.desc:key.keyboard.g
+key_key.arsmagicalegacy.next_shape_group:key.keyboard.c:SHIFT
+key_key.arsmagicalegacy.configure_spell:key.keyboard.m:CONTROL
 soundCategory_master:0.01986755
 soundCategory_music:0.0
 soundCategory_record:1.0
